lib/liquid_crypto: increase max HMAC key length to 1024
- Bump DAO_LC_MAX_AUTH_KEY_LEN from 128 to 1024
- Change auth_key_len from uint8_t to uint16_t

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpu-offload/+/162980
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>

+++ lib/liquid_crypto/liquid_crypto_sym.h
@@ -53,7 +53,7 @@ struct dao_lc_sym_sess_meta {
 	enum lc_crypto_op_type op_type;
 
 	/* HMAC authentication key length*/
-	uint8_t auth_key_len;
+	uint16_t auth_key_len;
 
 	/* HMAC authentication key */
 	uint8_t auth_key[DAO_LC_MAX_AUTH_KEY_LEN];
